address gemini comments

+++ clearnode/config/migrations/postgres/20251222000000_initial_schema.sql
@@ -6,7 +6,7 @@ CREATE TABLE channels (
     channel_id CHAR(66) PRIMARY KEY,
     user_wallet CHAR(42) NOT NULL,
     type SMALLINT NOT NULL, -- ChannelType enum: 0=void, 1=home, 2=escrow
-    blockchain_id INTEGER NOT NULL, -- uint64
+    blockchain_id BIGINT NOT NULL, -- uint64
     token CHAR(42) NOT NULL,
     challenge_duration BIGINT NOT NULL DEFAULT 0,
     challenge_expires_at TIMESTAMPTZ,
@@ -126,7 +126,7 @@ CREATE INDEX idx_app_ledger_v1_wallet ON app_ledger_v1(wallet);
 CREATE TABLE contract_events (
     id BIGSERIAL PRIMARY KEY,
     contract_address CHAR(42) NOT NULL,
-    blockchain_id INTEGER NOT NULL,
+    blockchain_id BIGINT NOT NULL,
     name VARCHAR(255) NOT NULL,
     block_number BIGINT NOT NULL,
     transaction_hash VARCHAR(255) NOT NULL,
@@ -141,7 +141,7 @@ CREATE TABLE blockchain_actions (
     id BIGSERIAL PRIMARY KEY,
     action_type SMALLINT NOT NULL,
     state_id CHAR(66),
-    blockchain_id INTEGER NOT NULL,
+    blockchain_id BIGINT NOT NULL,
     action_data JSONB,
     status SMALLINT NOT NULL DEFAULT 0,
     retry_count INTEGER NOT NULL DEFAULT 0,
